# webapp/plugins/identitystore/modules/quality/api/swagger/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Plugin identitystore-quality",
        "description": "Identitystore-quality API"
    },
    "produces": [
        "application/json"
    ],
    "paths": {
        "/suspiciousidentitys": {
            "get": {
                "description": "Get SuspiciousIdentity list",
                "tags": [
                    "SuspiciousIdentity"
                ],
                "responses": {
                    "200": {
                        "description": "Return the SuspiciousIdentity list",
                        "schema": {
                            "$ref": "#/definitions/SuspiciousIdentityListResult"
                        }
                    },
                    "204": {
                        "description": "No content",
                        "schema": {
                            "$ref": "#/definitions/NoContent"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create SuspiciousIdentity",
                "tags": [
                    "SuspiciousIdentity"
                ],
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "formData",
                        "description": "customer_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the SuspiciousIdentity created",
                        "schema": {
                            "$ref": "#/definitions/SuspiciousIdentityResult"
                        }
                    },
                    "400": {
                        "description": "Empty parameter",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/suspiciousidentitys/{id}": {
            "get": {
                "description": "Get SuspiciousIdentity",
                "tags": [
                    "SuspiciousIdentity"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the SuspiciousIdentity",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the SuspiciousIdentity",
                        "schema": {
                            "$ref": "#/definitions/SuspiciousIdentityResult"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Modify SuspiciousIdentity",
                "tags": [
                    "SuspiciousIdentity"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the SuspiciousIdentity to modify",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "customer_id",
                        "in": "formData",
                        "description": "customer_id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the SuspiciousIdentity modified",
                        "schema": {
                            "$ref": "#/definitions/SuspiciousIdentityResult"
                        }
                    },
                    "400": {
                        "description": "Empty parameter",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete SuspiciousIdentity",
                "tags": [
                    "SuspiciousIdentity"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the SuspiciousIdentity to delete",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the SuspiciousIdentity deleted (No Content)",
                        "schema": {
                            "$ref": "#/definitions/NoContent"
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "NoContent": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "SuspiciousIdentityListResult": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/SuspiciousIdentity"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "SuspiciousIdentityResult": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/SuspiciousIdentity"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "SuspiciousIdentity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "customer_id": {
                    "type": "string"
                }
            }
        }
    }
}